backend/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
certifi==2024.8.30
click==8.1.7
distlib==0.3.9
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
fastapi==0.115.2
filelock==3.16.1
gitdb==4.0.11
GitPython==3.1.41
greenlet==3.1.1
h11==0.14.0
idna==3.10
jose==1.0.0
Mako==1.3.6
MarkupSafe==3.0.2
packaging==24.1
passlib==1.7.4
pipenv==2024.1.0
platformdirs==4.3.6
psycopg2-binary==2.9.10
pyasn1==0.6.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-jose==3.3.0
rsa==4.9
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
virtualenv==20.26.6
